+++ flatlaf-testing/src/main/java/com/formdev/flatlaf/testing/FlatTestFrame.java
@@ -440,7 +440,7 @@ public class FlatTestFrame
 		sizeVariantComboBox.setVisible( visible );
 	}
 
-	private void updateComponentsRecur( Container container, BiConsumer<Component, String> action ) {
+	void updateComponentsRecur( Container container, BiConsumer<Component, String> action ) {
 		for( Component c : container.getComponents() ) {
 			if( c instanceof JPanel || c instanceof JDesktopPane ) {
 				updateComponentsRecur( (Container) c, action );